Non Latin characters in anchors do not encode well #54
Added option to include an anchor for committee members. The value will
be used in the committee member carousel to link to the committee member
page, and used as the ID on the committee member page. This field needs
to be pure Latin characters with no diacritics (e.g. accents, cedilla).
If characters w/ diacritics are included, they are encoded in the URL
and are less readable. Sample data was included in the committee.yml
file using a Kanji->Latin conversion tool.

Signed-off-by: Martin Lowe <martin.lowe@eclipse-foundation.org>